Fixed bug with os.path package not recognized
`astroid` had a special clause for "os.path", mapping it correctly to "ntpath", but to the "ntpath" in the executable environment, not the one in the Python source code we downloaded for the KG creation. This way we were missing the info we added to the ntpath node during preprocessing, and so missing the package individual and its related triples.